Move test formally
